src-backend/products/serializers.py
from rest_framework import serializers
from .models import Product, ProductAttachment

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'


class ProductAttachmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = ProductAttachment
        fields = '__all__'