src-backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
